tools/requirements.txt
markdown==3.5.1
python-docx
beautifulsoup4
